feat: generate map view methods
