+++ app/src/main/res/layout/activity_main.xml
@@ -7,69 +7,9 @@
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <LinearLayout
-        android:id="@+id/searchWindow"
+    <com.kakao.vectormap.MapView
+        android:id="@+id/map_view"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="20dp"
-        android:orientation="horizontal"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:layout_editor_absoluteX="10dp">
-
-        <EditText
-            android:id="@+id/searchInput"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:background="@android:color/transparent"
-            android:hint="@string/search_input_hint"
-            android:inputType="text" />
-
-        <ImageView
-            android:id="@+id/deleteInput"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:srcCompat="@drawable/baseline_clear_24" />
-    </LinearLayout>
-
-    <View
-        android:id="@+id/searchBar"
-        android:layout_width="match_parent"
-        android:layout_height="1.5dp"
-        android:layout_marginLeft="20dp"
-        android:layout_marginTop="3dp"
-        android:layout_marginRight="20dp"
-        android:layout_marginBottom="20dp"
-        android:background="#353535"
-        app:layout_constraintTop_toBottomOf="@id/searchWindow" />
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/no_searched_data"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/searchBar" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/searchHistory"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="20dp"
-        android:orientation="horizontal"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        app:layout_constraintTop_toBottomOf="@id/searchBar"
-        tools:listitem="@layout/search_history_module" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/placesRView"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginTop="20dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/searchHistory"
-        tools:listitem="@layout/place_module" />
+        android:layout_height="match_parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
